Share global read-only values using React Context
Since these values are read-only, it is safe to share them across
the application. Using React Context is a more ergonomic way than
passing them as props to every component, and ensures that they all
use the same instance of e.g. the store.

That said, because we want Panes to be independent of the Data
Browser, they do still get passed these values as props (for Panes
written in React) or regular arguments (for plain Javascript
Panes).