Add useAppOpener hook to seamlessly open an app
This hook provides a function that, when called with an origin and
the access modes it needs:

1. Gives that origin the required permissions.
2. Opens that app with the Pod server's URL as a query parameter
   (`idp` - the identify provider).

The opened app can then use that to start an authorisation flow
against that Pod server, which should be immediately successful
thanks to the new permissions, and then redirect back to the app
again. It will almost appear as if the app was just opened
directly, without the need to log in again.